[ST-11518] Delete whole attributes map instead of individual attributes (#75)
When logging out and deleting user attributes, we only deleted their
selection of attributes, but attributes themselves would still be left
over, resulting in duplicate attributes being shown when switching
between tenants.

In this PR, we also fix the issue where duplicate attributes can get
shown if app is opened as the attributes are being added / updated in
the CMS. In that case, we reset the old attributes before updating local
storage with the new ones. This should mostly affect us internally, it
will also affect users and clear their selection if we add or update new
attributes. We can add a better diff algorithm in the future for this.



https://github.com/stormideas/storyteller-showcase-ios-internal/assets/155957244/32df2505-3525-4bc6-8d6e-d5b8d318d37b